requirements.txt
django==3.2.5
django-notifications-hq==1.6.0
django-crispy-forms==1.12.0